requirements.txt
openai==1.3.7
pyTelegramBotAPI==4.13.0
python-dotenv
